requirements.txt
click==7.1.2
bcrypt>=3.1.3
Flask>=1.1.1
Flask-SQLAlchemy==2.4.4
Flask-BasicAuth>=0.2.0
Flask-Admin>=1.5.6
Flask-Bcrypt>=0.7.1
flask-cors>=3.0.9
flask-expects-json>=1.5.0
SQLAlchemy==1.3.23
PyJWT>=1.4.2
numpy>=1.20.1
beautifulsoup4==4.9.1
cachetools>=3.0.0
requests==2.24.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Werkzeug==1.0.1
